refactor(app): move the new-session flow state into ainb-app
The Configure form (ConfigureState, the branch picker, preset and base
selections, LaunchSpec) and the repo picker (PickRepoState, its rows,
favourites and auth status) move to ainb_app::components::new_session.
The renderers, the step dispatcher and both screens' key handlers stay in
ainb-core; the key handlers take crossterm KeyEvents until they move onto
Chord with the state machine.
