+++ ainb-tui/crates/ainb-app/src/components/new_session/mod.rs
@@ -1,0 +1,5 @@
+// ABOUTME: New-session flow state (`pick_repo` + `configure`). The two screens'
+// renderers, key handlers and the step dispatcher live in `ainb-core`.
+
+pub mod configure;
+pub mod pick_repo;
